trap #2 for a crash
git-svn-id: http://svn.miranda-ng.org/main/trunk@5652 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/TabSRMM/src/chat/chat.h
@@ -154,7 +154,7 @@ struct STATUSINFO
 {
 	TCHAR      *pszGroup;
 	HICON       hIcon;
-	WORD        Status;
+	DWORD       Status;
 	STATUSINFO *next;
 };
